refactor: move rpmbuild commands into own module
